chore(seeders): refactor seeders to resolve code climate issues
- create json data

[Finishes #164495531]
