+++ src/server/seeders/incidents.js
@@ -1,8 +1,8 @@
-const incidents = require('../utils/data/incidents');
+const incidents = require('../utils/data/incidents.json');
 
 module.exports = {
   up(queryInterface) {
-    return queryInterface.bulkInsert('Incidents', incidents);
+    return queryInterface.bulkInsert('Incidents', incidents.incidents);
   },
 
   down(queryInterface) {
