--- src/server/utils/data/incidents.js
@@ -1,257 +1,0 @@
-const incidents = [
-  {
-    id: 'cjfkubrlv0001tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0002tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0003tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0004tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0005tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0006tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0007tgxs3mre',
-    subject: 'Abusive Language',
-    description: 'Someone said some things to me I cannot even type',
-    dateOccurred: '12-10-2017',
-    levelId: '1',
-    statusId: '2',
-    categoryId: '22',
-    locationId: 'cjee256gt0000ioxs69v4870x',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cjfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c1jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c2jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c3jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c4jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c5jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'c6jfkubrlv0001tgxs3',
-    subject: 'Theft',
-    description: 'Someone picked up my phone from the table',
-    dateOccurred: '3-1-2018',
-    levelId: '2',
-    statusId: '1',
-    categoryId: '23',
-    locationId: 'cjee24n0n0000hfxsefer9tjh',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-
-  {
-    id: 'cjfkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj1fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj2fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj3fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj4fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj5fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-  {
-    id: 'cj6fkubrlv0001tsjksuis3',
-    subject: 'Discrimination',
-    description: 'Someone said something that was discriminative to my gender',
-    dateOccurred: '2-3-2018',
-    levelId: '3',
-    statusId: '3',
-    categoryId: '24',
-    locationId: 'cjee241h20000g7xsfzd572sl',
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-  },
-];
-
-module.exports = incidents;
